+++ integration_tests/suite/helpers/fixtures.py
@@ -280,6 +280,8 @@ def user(**user_args):
         user_args['firstname'] = _random_string(20)
     if 'lastname' not in user_args:
         user_args['lastname'] = _random_string(20)
+    if 'purpose' not in user_args:
+        user_args['purpose'] = 'user'
     if 'enabled' not in user_args:
         user_args['enabled'] = True