+++ wazo_auth/plugins/http/users/schemas.py
@@ -6,23 +6,27 @@ from xivo.mallow import validate
 from wazo_auth.schemas import BaseSchema
 
 
-class UserPostSchema(BaseSchema):
+class _BaseUserSchema(BaseSchema):
+
+    username = fields.String(validate=validate.Length(min=1, max=256), required=True)
+    firstname = fields.String(missing=None, allow_none=True)
+    lastname = fields.String(missing=None, allow_none=True)
+    purpose = fields.String(
+        missing='user',
+        validate=validate.OneOf(['user', 'internal', 'external_api'])
+    )
+    enabled = fields.Boolean(missing=True)
+
+
+class UserPostSchema(_BaseUserSchema):
 
     uuid = fields.UUID()
-    username = fields.String(validate=validate.Length(min=1, max=256), required=True)
     password = fields.String(validate=validate.Length(min=1), allow_none=True)
-    firstname = fields.String(missing=None, allow_none=True)
-    lastname = fields.String(missing=None, allow_none=True)
     email_address = fields.Email(allow_none=True)
-    enabled = fields.Boolean(missing=True)
 
 
-class UserPutSchema(BaseSchema):
-
-    username = fields.String(validate=validate.Length(min=1, max=256), required=True)
-    firstname = fields.String(missing=None, allow_none=True)
-    lastname = fields.String(missing=None, allow_none=True)
-    enabled = fields.Boolean(missing=True)
+class UserPutSchema(_BaseUserSchema):
+    pass
 
 
 class ChangePasswordSchema(BaseSchema):